db: переезд с Neon (Postgres) на MariaDB на Hostinger
Neon остановил проект 13.09 по лимиту трафика бесплатного плана, приложение
лежало. База теперь на том же Hostinger, что и сайт.

- lib/schema.sql — консолидированная схема MariaDB: те же таблицы, CHECK с
  прежними именами, внешние ключи, индексы; utf8mb4_nopad_bin, DATETIME(6) UTC
- lib/db.ts — пул mysql2 (5 соединений, wait_timeout сессии 8 ч: Hostinger
  рвёт простой через 20 с и даёт 500 подключений в час), строгий режим,
  ANSI_QUOTES и PIPES_AS_CONCAT; BOOLEAN → true/false, DATE → как у pg
- lib/db-format.ts — сборка текста запроса с экранированием (+ тесты)
- 266 запросов переписаны: ON DUPLICATE KEY, UPDATE…RETURNING → affectedRows/
  SELECT, DISTINCT ON → ROW_NUMBER, NULLS LAST, ILIKE, регулярки, TO_BASE64/
  UNHEX, INTERVAL, NOW(6), "key"; время в DATETIME передаётся объектом Date
- middleware на runtime nodejs (mysql2 не работает в edge)
- scripts/pg-dump-to-mysql.mjs — перенос из дампа Neon (сверено: строки всех
  17 таблиц и md5 всех 358 файлов)
- scripts/mysql-to-pg.mjs — ОТКАТ: данные обратно в Postgres (проверено на
  PGlite со схемой pre-mysql); код отката — метка pre-mysql
- scripts/check-sql.mjs (npm run db:check) — PREPARE каждого запроса на базе
- backup.yml — ежедневный mariadb-dump вместо pg_dump

+++ lib/tour.ts
@@ -82,9 +82,9 @@ export async function tourSteps(user: CurrentUser | null, locale: Locale): Promi
     // trucks.id <> 1 — посеянная заглушка «Трак не настроен» лежит в схеме под
     // первым номером и есть в любой установке; настоящий трак получает следующий.
     sql`SELECT
-          (SELECT count(*)::int FROM users  WHERE is_demo = FALSE)                     AS users,
-          (SELECT count(*)::int FROM trucks WHERE company_id = 'default' AND id <> 1)  AS trucks,
-          (SELECT count(*)::int FROM loads  WHERE company_id = 'default')              AS loads`,
+          (SELECT count(*) FROM users  WHERE is_demo = FALSE)                     AS users,
+          (SELECT count(*) FROM trucks WHERE company_id = 'default' AND id <> 1)  AS trucks,
+          (SELECT count(*) FROM loads  WHERE company_id = 'default')              AS loads`,
   ])
   const n = (counts as { users: number; trucks: number; loads: number }[])[0] ?? {
     users: 1,
